@@ -12,28 +12,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^2.3.1",
-    "@angular/compiler": "^2.3.1",
-    "@angular/core": "^2.3.1",
-    "@angular/forms": "^2.3.1",
-    "@angular/http": "^2.3.1",
-    "@angular/platform-browser": "^2.3.1",
-    "@angular/platform-browser-dynamic": "^2.3.1",
-    "@angular/router": "^3.3.1",
-    "@types/xlsx": "0.0.36",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^8.1.2",
+    "@angular/cdk": "^8.1.1",
+    "@angular/cli": "8.3.15",
+    "@angular/common": "~8.1.0",
+    "@angular/compiler": "~8.1.0",
+    "@angular/core": "~8.1.0",
+    "@angular/forms": "~8.1.0",
+    "@angular/material": "^8.1.1",
+    "@angular/platform-browser": "~8.1.0",
+    "@angular/platform-browser-dynamic": "~8.1.0",
+    "@angular/router": "~8.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.0",
+    "angular5-csv": "0.2.11",
+    "bootstrap": "^4.3.1",
     "export-to-csv": "0.2.1",
     "file-saver": "2.0.2",
-    "ngx-bootstrap": "5.2.0",
-    "rxjs": "^5.0.1",
-    "ts-helpers": "^1.1.1",
+    "hammerjs": "^2.0.8",
+    "httpclient": "0.1.0",
+    "jquery": "^3.3.1",
+    "popper.js": "^1.14.3",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.9.0",
     "xlsx": "0.15.1",
-    "zone.js": "^0.7.2"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular/compiler-cli": "^2.3.1",
     "@types/jasmine": "2.5.38",
-    "@types/node": "^6.0.42",
+    "@types/node": "6.14.9",
     "angular-cli": "1.0.0-beta.28.3",
     "codelyzer": "~2.0.0-beta.1",
     "jasmine-core": "2.5.2",
