@@ -1,22 +1,42 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { FormsModule, ReactiveFormsModule} from '@angular/forms';
+import {HttpClientModule} from '@angular/common/http';
+
+import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
+import {MatNativeDateModule} from '@angular/material/core';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+// import {DemoMaterialModule} from './app/material-module';
+import { MatTableModule } from '@angular/material';
 
 import { AppComponent } from './app.component';
+import { TableComponent } from './table/table.component';
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    
+    TableComponent,
+    BrowserModule,
+    MatNativeDateModule,
+
+
   ],
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule
+    BrowserAnimationsModule,
+    ReactiveFormsModule,
+    // DemoMaterialModule,
+    HttpClientModule,
+    MatTableModule
+
   ],
   providers: [],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
+  exports:[MatTableModule]
 })
 export class AppModule { }
+
+platformBrowserDynamic().bootstrapModule(AppModule)
+  .catch(err => console.error(err));
